+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 Starting from v2.2.5, all notable changes to this project will be documented in this file.
 
+## v2.12.2
+
+- Always use fallback locale if the preferred locale is not available
+- Optimize the serial port reading process
+
 ## v2.12.1
 
 - Sort earthquake event source API by name in frontend